requirements.txt
Flask==3.1.0
groq
python-dotenv
beautifulsoup4
